edit loggers registry, return logging functions in parser.py
