+++ oml/lightning/pipelines/parser.py
@@ -1,16 +1,22 @@
+import os
+import warnings
 from pathlib import Path
-from typing import Any, Dict, Optional
+from typing import Any, Dict, Optional, Union
 
+import albumentations as albu
 import torch
 from pytorch_lightning.callbacks import ModelCheckpoint
+from pytorch_lightning.loggers import LightningLoggerBase, NeptuneLogger, WandbLogger
 from pytorch_lightning.plugins import DDPPlugin
 
-from oml.const import TCfg
+from oml.const import PROJECT_ROOT, TCfg
 from oml.datasets.base import DatasetWithLabels
 from oml.interfaces.samplers import IBatchSampler
+from oml.registry.loggers import get_logger_by_cfg
 from oml.registry.samplers import SAMPLERS_CATEGORIES_BASED, get_sampler_by_cfg
 from oml.registry.schedulers import get_scheduler_by_cfg
-from oml.utils.misc import dictconfig_to_dict
+from oml.registry.transforms import get_transforms_by_cfg
+from oml.utils.misc import dictconfig_to_dict, flatten_dict
 
 
 def parse_engine_params_from_config(cfg: TCfg) -> Dict[str, Any]:
@@ -55,6 +61,57 @@ def check_is_config_for_ddp(cfg: TCfg) -> bool:
     return bool(cfg["strategy"])
 
 
+def initialize_logging(cfg: TCfg) -> LightningLoggerBase:
+    """
+    Logger initialisation.
+
+    If Neptune Logger is used, we also upload to its cloud some files
+    which are good to have for reproducibility.
+
+    """
+    cfg = dictconfig_to_dict(cfg)
+
+    logger = get_logger_by_cfg(cfg["logger"])
+
+    dict_to_log = {**cfg, **{"dir": Path.cwd()}}
+    logger.log_hyperparams(flatten_dict(dict_to_log, sep="|"))
+
+    if isinstance(logger, NeptuneLogger):
+        warnings.warn(
+            "Unfortunately, in the case of using Neptune, you may experience that long experiments are"
+            "stacked and not responding. It's not an issue on OML's side, so, we cannot fix it. You can use"
+            "Tensorboard logger instead, for this simply leave <NEPTUNE_API_TOKEN> unfilled."
+        )
+        upload_files_to_neptune_cloud(logger, cfg)
+
+    return logger
+
+
+def upload_files_to_neptune_cloud(logger: NeptuneLogger, cfg: TCfg) -> None:
+    assert isinstance(logger, NeptuneLogger)
+
+    # save transforms as files
+    for key, val in cfg.items():
+        if "transforms" in key:
+            try:
+                transforms = get_transforms_by_cfg(cfg[key])
+                if isinstance(transforms, albu.Compose):
+                    transforms_file = str(Path(".hydra/") / f"{key}.yaml") if Path(".hydra").exists() else f"{key}.yaml"
+                    albu.save(filepath=transforms_file, transform=transforms, data_format="yaml")
+                    logger.run[key].upload(str(transforms_file))
+            except Exception:
+                print(f"We are not able to interpret {key} as albumentations transforms and log them as a file.")
+
+    # log source code
+    source_files = list(map(lambda x: str(x), PROJECT_ROOT.glob("**/*.py"))) + list(
+        map(lambda x: str(x), PROJECT_ROOT.glob("**/*.yaml"))
+    )
+    logger.run["code"].upload_files(source_files)
+
+    # log dataset
+    logger.run["dataset"].upload(str(Path(cfg["dataset_root"]) / cfg["dataframe_name"]))
+
+
 def parse_scheduler_from_config(cfg: TCfg, optimizer: torch.optim.Optimizer) -> Dict[str, Any]:
     if cfg.get("scheduling"):
         scheduler_kwargs = {
@@ -96,6 +153,8 @@ def parse_ckpt_callback_from_config(cfg: TCfg) -> ModelCheckpoint:
 __all__ = [
     "parse_engine_params_from_config",
     "check_is_config_for_ddp",
+    "initialize_logging",
+    "upload_files_to_neptune_cloud",
     "parse_scheduler_from_config",
     "parse_sampler_from_config",
     "parse_ckpt_callback_from_config",
